refactor: moving dataset revision to tasks
